pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rdb"
version = "0.1.0"
description = "Retrieval Database for Arch Wiki documentation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
   {name = "Topher Ludlow", email = "topherludlow@protonmail.com"}
]
classifiers = [
   "Development Status :: 3 - Alpha",
   "Intended Audience :: Developers",
   "License :: OSI Approved :: MIT License",
   "Operating System :: OS Independent",
   "Programming Language :: Python :: 3",
   "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
]
dependencies = [
   "requests>=2.31.0",
   "beautifulsoup4>=4.12.0",
   "lxml>=4.9.0",
   "numpy>=1.24.0",
   "pandas>=2.0.0",
   "torch>=2.0.0",
   "transformers>=4.30.0",
   "sentence-transformers>=2.2.0",
   "faiss-cpu>=1.7.4",
   "tqdm>=4.65.0",
   "python-dotenv>=1.0.0",
   "click>=8.1.0",
   "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
   "pytest>=7.4.0",
   "pytest-cov>=4.1.0",
   "black>=23.0.0",
   "flake8>=6.0.0",
   "isort>=5.12.0",
]
gpu = [
   "faiss-gpu>=1.7.4",
]

[project.scripts]
rdb = "cli.main:main"

[project.urls]
Homepage = "https://github.com/Topher2014/rdb"
Repository = "https://github.com/Topher2014/rdb"
Issues = "https://github.com/Topher2014/rdb/issues"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
